+++ packages/i18n/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @orderly.network/i18n
 
+## 2.12.3
+
+### Patch Changes
+
+- publish
+- 528e3f4: publish
+
+## 2.12.3-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 2.12.2
 
 ### Patch Changes